test(playwright-ct): break up pt-input tests to separate stories
Having all the tests in one story is getting quite unmanageable.
I have broken up all the tests for the pt-input into own stories
about topics like DecoratorsStory, StylesStory, AnnotationStory etc.

This allows us to write smaller and more specialized schemas for each
story and it will also be much easier to add tests without having
to deal with complex schemas and specs.
