+++ packages/sanity/src/core/form/__workshop__/index.ts
@@ -15,17 +15,5 @@ export default defineScope({
       title: 'Form builder (unfinished)',
       component: lazy(() => import('./FormBuilderStory')),
     },
-    {
-      name: 'form-builder-pte-input',
-      title: 'Form builder: Portable Text Input',
-      component: lazy(
-        () => import('../../../../playwright-ct/tests/formBuilder/PortableTextInputStory'),
-      ),
-    },
-    {
-      name: 'form-builder-array-input',
-      title: 'Form builder: Array Input',
-      component: lazy(() => import('../../../../playwright-ct/tests/formBuilder/ArrayInputStory')),
-    },
   ],
 })
